smartpath-ai/requirements.txt
fastapi
uvicorn
pandas
streamlit
sentence-transformers
torch
requests
PyMuPDF
google-generativeai
